+++ src/main/factories/controllers/signup/signup-controller-factory.ts
@@ -1,0 +1,11 @@
+import { SignUpController } from '../../../../presentation/controllers/signUp/signup-controller'
+import { Controller } from '../../../../presentation/protocols'
+import { makeSignUpValidation } from './signup-validation-factory'
+import { makeDbAuthentication } from '../../usecases/authentication/db-authentication-factory'
+import { makeDbAddAccount } from '../../usecases/add-account/db-add-account-factory'
+import { makeLogControllerDecorator } from '../../decorators/log-controller-decorator-factory'
+
+export const makeSignUpController = (): Controller => {
+  const signUpController = new SignUpController(makeDbAddAccount(), makeSignUpValidation(), makeDbAuthentication())
+  return makeLogControllerDecorator(signUpController)
+}
